+++ runtime/test/unit/ujima/importer_test.clj
@@ -21,13 +21,13 @@
 (deftest a-valid-file-applies-across-scopes
   (let [dir (fresh!)
         result (importer/import!
-                 [{:scope :device :setting [:system :hostname] :value "shule-3"}
+                 [{:scope :device :setting [:system :name] :value "shule-3"}
                   {:scope :device :setting [:system :timezone] :value "Africa/Nairobi"}
                   {:scope :circle :setting [:circle :name]     :value "Shule Circle"}]
                  {})]
     (is (:ok? result))
     (is (= 3 (:applied result)))
-    (is (= "shule-3" (value [:system :hostname])))
+    (is (= "shule-3" (value [:system :name])))
     (is (= "Shule Circle" (value [:circle :name])))
     (let [raw (edn/read-string (slurp (str dir "/device.edn")))]
       (is (= defs/schema (:schema raw)) "written through control's writer, schema-stamped"))))
@@ -36,20 +36,20 @@
 (deftest one-bad-entry-applies-nothing
   (let [dir (fresh!)
         result (importer/import!
-                 [{:scope :device :setting [:system :hostname] :value "ok-name"}
+                 [{:scope :device :setting [:system :name] :value "ok-name"}
                   {:scope :device :setting [:not :a :setting]  :value 1}]
                  {})]
     (is (not (:ok? result)))
     (is (= 0 (:applied result)))
     (is (= 1 (count (:errors result))))
     (is (not (fs/exists? (str dir "/device.edn"))) "all-or-nothing: no scope file written")
-    (is (nil? (value [:system :hostname])))))
+    (is (nil? (value [:system :name])))))
 
 
 (deftest scope-not-allowed-is-an-error-not-a-silent-prune
   (fresh!)
   (let [result (importer/import!
-                 [{:scope :circle :setting [:system :hostname] :value "x"}]
+                 [{:scope :circle :setting [:system :name] :value "x"}]
                  {})]
     (is (not (:ok? result)))
     (is (re-find #"cannot be set in scope" (-> result :errors first :error)))))
@@ -58,16 +58,16 @@
 (deftest shape-errors-speak-the-schema-language
   (fresh!)
   (let [result (importer/import!
-                 [{:scope :device :setting [:system :hostname] :value "no spaces!"}]
+                 [{:scope :device :setting [:system :name] :value "no spaces!"}]
                  {})]
     (is (not (:ok? result)))
-    (is (re-find #"hostname" (-> result :errors first :error)))))
+    (is (re-find #"name" (-> result :errors first :error)))))
 
 
 (deftest unknown-scope-is-an-error
   (fresh!)
   (let [result (importer/import!
-                 [{:scope :nope :setting [:system :hostname] :value "x"}]
+                 [{:scope :nope :setting [:system :name] :value "x"}]
                  {})]
     (is (not (:ok? result)))
     (is (re-find #"unknown scope" (-> result :errors first :error)))))
@@ -87,12 +87,12 @@
 (deftest validate-only-touches-nothing
   (let [dir (fresh!)
         result (importer/import!
-                 [{:scope :device :setting [:system :hostname] :value "shule-3"}]
+                 [{:scope :device :setting [:system :name] :value "shule-3"}]
                  {:validate-only true})]
     (is (:ok? result))
     (is (= 0 (:applied result)))
     (is (not (fs/exists? (str dir "/device.edn"))))
-    (is (nil? (value [:system :hostname])))))
+    (is (nil? (value [:system :name])))))
 
 
 (deftest an-empty-or-non-vector-file-is-an-error